Fix extra-args handling (#9)
* Fix extra-args handling

The current code doesn't actually pass the extra args to nox.
The action echos the command about to be run before running it, and #6
only updates the `echo` part, not the actual command being run.

Also, it would be good to also pass the args when installing the session
dependencies as some sessions may make decisions about what to install
based on the arguments passed (I've written code like this before, at
least).

Fixes: https://github.com/ansible-community/github-action-run-nox/pull/6

* Add ci test for extra-args functionality